pyproject.toml
[project]
name = "thesis-dgp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = ['pyro-ppl', 'matplotlib','seaborn','pandas','scikit-learn', 'polars',]
